+++ app/src/main/kotlin/com/jainamj/myapplication/data/db/DbHelper.kt
@@ -6,7 +6,7 @@ import io.reactivex.Observable
 interface DbHelper {
     fun addAllUsers(users: List<User>): Observable<Boolean>
 
-    fun addUser(user: User): Observable<Int>
+    fun addUser(user: User): Observable<Long>
 
     fun updateUser(user: User): Observable<Boolean>
